+++ database/migrations/2020_01_18_154330_create_transactions_table.php
@@ -16,8 +16,8 @@ class CreateTransactionsTable extends Migration
         Schema::create('transactions', function (Blueprint $table) {
             $table->bigIncrements('id');
             $table->unsignedInteger('quantity');
-            $table->unsignedInteger('buyer_id');
-            $table->unsignedInteger('product_id');
+            $table->unsignedBigInteger('buyer_id');
+            $table->unsignedBigInteger('product_id');
             $table->timestamps();
             $table->foreign('buyer_id')->references('id')->on('users');
             $table->foreign('product_id')->references('id')->on('products');
