+++ database/migrations/2020_01_18_165858_category_product_table.php
@@ -14,11 +14,10 @@ class CategoryProductTable extends Migration
     public function up()
     {
         Schema::create('category_product', function (Blueprint $table) {
-            $table->unsignedInteger('category_id');
-            $table->unsignedInteger('product_id');
+            $table->unsignedBigInteger('category_id');
+            $table->unsignedBigInteger('product_id');
             $table->foreign('category_id')->references('id')->on('categories');
             $table->foreign('product_id')->references('id')->on('products');
-
         });
     }
